Move `Helper` out of `ModelAnnotator`
