+++ lib/annotate_rb/model_annotator.rb
@@ -3,7 +3,6 @@
 module AnnotateRb
   module ModelAnnotator
     autoload :Annotator, 'annotate_rb/model_annotator/annotator'
-    autoload :Helper, 'annotate_rb/model_annotator/helper'
     autoload :FilePatterns, 'annotate_rb/model_annotator/file_patterns'
     autoload :Constants, 'annotate_rb/model_annotator/constants'
     autoload :PatternGetter, 'annotate_rb/model_annotator/pattern_getter'
