--- spec/lib/annotate_rb/model_annotator/helper_spec.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe AnnotateRb::ModelAnnotator::Helper do
-end
